Move status and statushistory commands to own package.
